feat: auto rerun
