+++ arex-schedule-web-api/src/main/java/com/arextest/schedule/beans/EventBusConfiguration.java
@@ -1,6 +1,5 @@
 package com.arextest.schedule.beans;
 
-import com.arextest.schedule.eventBus.PlanAutoRerunEventListener;
 import com.google.common.eventbus.AsyncEventBus;
 import java.util.concurrent.ExecutorService;
 import org.springframework.context.annotation.Bean;
@@ -14,7 +13,7 @@ import org.springframework.context.annotation.Configuration;
 public class EventBusConfiguration {
 
   @Bean
-  public AsyncEventBus asyncEventBus(ExecutorService autoRerunExecutorService) {
+  public AsyncEventBus autoRerunAsyncEventBus(ExecutorService autoRerunExecutorService) {
     return new AsyncEventBus("replay-event-bus", autoRerunExecutorService);
   }
