--- arex-schedule-web-api/src/main/java/com/arextest/schedule/eventBus/EventBusUtils.java
@@ -1,10 +1,0 @@
-package com.arextest.schedule.eventBus;
-
-/**
- * @author wildeslam.
- * @create 2023/12/11 16:28
- */
-public class EventBusUtils {
-
-
-}
